backend/requirements.txt
asgiref==3.8.1
autoenv==1.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.2
django-ckeditor-5==0.2.11
django-cors-headers==4.7.0
django-js-asset==3.1.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
Flask==3.1.0
fsspec==2025.7.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
mysql-connector-python==8.0.30
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pytesseract==0.3.10
opencv-python==4.9.0.80
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
reportlab==4.0.4
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.30.0
Werkzeug==3.1.3
